AudioCD ioslave is now a plugin system!
- Coping what seemed right from k3b I have taken the plugins and put them each
into their own library. (This by itself was very little work as they were
already abstracted classes).  Really the only biggest change/addition was the
addition of the 40 LOC find_plugins() function in the base class which opens
the libraries and loads the encoders.  In the future much closer intigration
with k3b is the goal.

-Second I moved the configure widget for each encoder into the encoder plugin
from the kcm.  This allows for the kcm to use find_plugins() and then load
their configure dialogs.  One of the really cool side effect for example is
that if lame sees that there is no lame library to run it wont return a
encoder to configure (and thus confusing the user).

-Third (tied very close to #2) was I converted the encoders to use KConfigXT.
I created kcfg files, modified the ui files, modified the code and added the
ability for the kcm to handle it (several three line for loops).  Because of
this change quite a lot of code that KConfigXT handles was able to be
removed from the kcm widget. On a side note k3b should do this.

Planned:
1) Incorperate k3b's Encoder from command line plugin
2) Ditch 50% of KAudioCreator and just use this new system.
3) Help port other apps to use this (k3b?,juk?,kcd?)

svn path=/trunk/kdemultimedia/kioslave/audiocd/; revision=297461