Somewhere along the way, exactly half of the text describing the autosearch functionality was changed to the logically opposite meaning. Nowadays, the configure option is 'Specify device' as opposed to 'Autosearch'. This makes the code for enabling the textbox simpler, sure.
Since the logical sense has changed, we need extra NOTs in the load and save methods.

Since the meaning of the option has changed, the whatsthis should describe what the option does, and not what it used to do. This would be a real problem if anyone actually read whatsthis text.

CCMAIL: kde@freebsd.org
CCMAIL: kde-multimedia@kde.org
CCMAIL: kde-i18n-doc@kde.org

PS. Docs people get an extra *grovel* in order to fix this totally wrong string. If it's out of the question to do now, drop me a note and I'll revert the whatsthis change ASAP.

svn path=/branches/KDE/3.5/kdemultimedia/kioslave/audiocd/; revision=494647