For FreeBSD, cdparanoia's autosearch combined with either ATAPICAM (ATAPI-SCSI) or plain ATAPI doesn't fill in the expected fields in the CAM structure; that causes crashes. So go case-by case and use what makes sense. For the ATAPI case, assume the user has specified a device name because otherwise we need to map an open fd back to a pathname (knowing it's in /dev/ helps, and it's just a bunch of fstat and stat calls, but still).
CCMAIL: kde@freebsd.org

PS. Markus Brueffer has patches to cdparanoia to add the Linux-style field cdda_device_name so that we can drop this whole chunk of ifdefs and just use that field. This moves the smarts into cdparanoia itself. We will add a configure check when that comes around.

svn path=/branches/KDE/3.5/kdemultimedia/kioslave/audiocd/; revision=494661