audiocd-kio: Compile vorbis support when detected.
Apply a patch from Felix Tiede to re-enable vorbis plugin support for
the audio CD KIOSlave. (It can be verified that support was not
previously compiled by running kcmshell4 audiocd).

Felix has ideas for further improvement but I wanted to ensure that KDE
SC 4.9.2 contained the fix. Thanks for the bug report, issue
investigation, and fix(es)!

FIXED-IN:4.9.2
BUG:307343