generate and use a local cdda_interface headerfile copy
Summary:
Generates and uses a local copy of the cdda_interface.h header file that
works with C++ instead of using the previous workaround based on
preprocessor macros. That was a hack that breaks the build when
cdda_interface.h includes other header files that use the `private`
keyword.

Test Plan: works as intended.

Reviewers: #frameworks, dfaure

Reviewed By: dfaure

Subscribers: dfaure

Differential Revision: https://phabricator.kde.org/D7448