Fix various issues with the device action
The predicate didn't work for mixed data CD's, so switch it to using a mask
instead so the action shows up. Secondly, change the icon to something more
fitting along with the name. This way the user isn't presented with *two*
"Open with File Manager" device actions if the CD happens to have both audio
and data.